New folder structure and smaller files
