+++ app.py
@@ -8,657 +8,22 @@
 # - Selector para preguntar a Drive, Dropbox o ambos.
 # ─────────────────────────────────────────────────────────────────────────────
 
-import os, io, json, time, logging, unicodedata, re
-from datetime import datetime
+# Main imports
+import os
 
-import numpy as np
-import streamlit as st
-import faiss
-
-from dotenv import load_dotenv
-from PyPDF2 import PdfReader
-
-# Google APIs
-from googleapiclient.discovery import build
-from googleapiclient.errors import HttpError
-from google_auth_oauthlib.flow import InstalledAppFlow, Flow
-from google.oauth2.credentials import Credentials
-
-# Dropbox SDK
-import dropbox
-from dropbox.exceptions import ApiError
-
-# LangChain v0.2
-from langchain_community.vectorstores import FAISS
-from langchain_community.docstore.in_memory import InMemoryDocstore
-from langchain_text_splitters import RecursiveCharacterTextSplitter
-from langchain_core.documents import Document
-from langchain_openai import OpenAIEmbeddings, ChatOpenAI
-# from langchain.schema import Document
+from langchain_openai import ChatOpenAI, OpenAIEmbeddings
 from langchain_core.messages import SystemMessage, HumanMessage
 
-# One drive
-import requests
-import msal
-import traceback
-
-# ─────────────────────────────────────────────────────────────────────────────
-# CONFIG
-# ─────────────────────────────────────────────────────────────────────────────
-
-load_dotenv(override=True)
-
-CLIENT_SECRET_FILE = "client_secret.json"
-SCOPES = ["https://www.googleapis.com/auth/drive.readonly"]
-
-OPENAI_API_KEY = (os.getenv("OPENAI_API_KEY") or "").strip().strip('"').strip("'")
-FOLDER_ID      = (os.getenv("FOLDER_ID") or "").strip().strip('"').strip("'")
-REDIRECT_URI   = (os.getenv("REDIRECT_URI") or "").strip().strip('"').strip("'")
-
-# Dropbox
-DROPBOX_APP_KEY    = (os.getenv("DROPBOX_APP_KEY") or "").strip().strip('"').strip("'")
-DROPBOX_APP_SECRET = (os.getenv("DROPBOX_APP_SECRET") or "").strip().strip('"').strip("'")
-DROPBOX_ROOT       = (os.getenv("DROPBOX_ROOT") or "/").strip()  # carpeta a indexar
-
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-LOGO_IN2AI = os.path.join(BASE_DIR, "img", "in2ai.png")
-LOGO_IGAPE = os.path.join(BASE_DIR, "img", "igape.png")
-
-if not OPENAI_API_KEY:
-    st.error("❌ Falta OPENAI_API_KEY en .env"); st.stop()
-if not FOLDER_ID:
-    st.error("❌ Falta FOLDER_ID en .env"); st.stop()
-if not REDIRECT_URI:
-    st.error("❌ Falta REDIRECT_URI en .env"); st.stop()
-
-logging.getLogger("googleapiclient.discovery_cache").setLevel(logging.ERROR)
-st.set_page_config(page_title="RAG Drive + Dropbox ACL", page_icon="💬", layout="wide")
-
-# OneDrive (Microsoft Graph)
-ONEDRIVE_CLIENT_ID = (os.getenv("ONEDRIVE_CLIENT_ID") or "").strip()
-ONEDRIVE_TENANT_ID = (os.getenv("ONEDRIVE_TENANT_ID") or "").strip()
-ONEDRIVE_AUTHORITY = f"https://login.microsoftonline.com/{ONEDRIVE_TENANT_ID}"
-ONEDRIVE_SCOPES = ["Files.Read.All", "User.Read"]
-ONEDRIVE_ROOT = (os.getenv("ONEDRIVE_ROOT") or "").strip()   # p.ej. "", "/RAG"
-GRAPH = "https://graph.microsoft.com/v1.0"
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# ESTILOS
-# ─────────────────────────────────────────────────────────────────────────────
-
-st.markdown("""
-<style>
-.topbar{display:flex;justify-content:space-between;align-items:center;
-padding:6px 10px;border-bottom:1px solid #eaeaea;margin-bottom:8px;}
-.topbar img{height:40px;}
-.chat-bubble-user{
-  background:#dcf8c6;border-radius:20px;padding:10px 14px;display:inline-block;
-  margin:6px 0;max-width:80%;
-}
-.chat-bubble-bot{
-  background:#ffffff;border:1px solid #eee;border-radius:20px;padding:10px 14px;
-  display:inline-block;margin:6px 0;max-width:80%;
-}
-.chat-row{display:flex;align-items:flex-start;}
-.chat-row.user{justify-content:flex-end;}
-.chat-row.bot{justify-content:flex-start;}
-section[data-testid="stSidebar"] .block-container { display:flex; flex-direction:column; align-items:center; text-align:center; }
-section[data-testid="stSidebar"] img { display:block; margin-left:auto; margin-right:auto; }
-section[data-testid="stSidebar"] .stButton>button { width:100%; max-width:240px; margin:0 auto; }
-</style>
-""", unsafe_allow_html=True)
-
-# ─────────────────────────────────────────────────────────────────────────────
-# HELPERS COMUNES
-# ─────────────────────────────────────────────────────────────────────────────
-
-def safe_execute(request, retries=6, backoff=1.7):
-    for i in range(retries):
-        try:
-            return request.execute()
-        except HttpError as e:
-            if getattr(e, "resp", None) and e.resp.status in (500, 502, 503, 504):
-                time.sleep(backoff ** i)
-                continue
-            raise
-    raise RuntimeError("Google API: demasiados fallos consecutivos (5xx).")
-
-def _cosine(a, b):
-    a = np.array(a); b = np.array(b)
-    denom = (np.linalg.norm(a) * np.linalg.norm(b)) + 1e-12
-    return float(np.dot(a, b) / denom)
-
-# ─────────────────────────────────────────────────────────────────────────────
-# GOOGLE DRIVE
-# ─────────────────────────────────────────────────────────────────────────────
-
-def oauth_login_drive():
-    """
-    Web OAuth para Drive (robusto en Streamlit + Docker):
-    - Si vuelve ?code=..., se intercambia aunque se haya perdido session_state.
-    - Usa los scopes EXACTOS devueltos en la URL (?scope=...) para evitar
-      "Scope has changed...".
-    - Si aún no hay code, muestra la URL de autorización.
-    - Tras éxito: guarda service en sesión, limpia la URL y devuelve service.
-    """
-    if not os.path.exists(CLIENT_SECRET_FILE):
-        st.error(f"❌ No existe {CLIENT_SECRET_FILE}. Revisa el client web en Google Cloud.")
-        st.stop()
-    if not REDIRECT_URI:
-        st.error("❌ Falta REDIRECT_URI en .env (debe coincidir EXACTAMENTE con la registrada).")
-        st.stop()
-
-    # ---------- Helpers (definidos DENTRO para que siempre existan) ----------
-    def _get_qp():
-        # Streamlit 1.33+ (st.query_params) y compatibilidad con anteriores
-        try:
-            qp = st.query_params
-            if hasattr(qp, "items"):
-                return {k: v for k, v in qp.items()}
-        except Exception:
-            pass
-        try:
-            return st.experimental_get_query_params() or {}
-        except Exception:
-            return {}
-
-    def _clear_qp():
-        # Borra ?code,&state,&scope de la URL para evitar reintentos
-        try:
-            st.query_params.update({})
-            return
-        except Exception:
-            pass
-        try:
-            st.experimental_set_query_params()
-        except Exception:
-            pass
-
-    def _first(v):
-        return v[0] if isinstance(v, list) else v
-
-    # ------------------------- Lógica principal ------------------------------
-    qp = _get_qp()
-    incoming_code   = _first(qp.get("code"))   if qp.get("code")   else None
-    incoming_state  = _first(qp.get("state"))  if qp.get("state")  else None
-    incoming_scope  = _first(qp.get("scope"))  if qp.get("scope")  else None
-    scopes_callback = [s for s in (incoming_scope or "").split() if s] or None
-
-    # 1) Si ya tenemos ?code=..., intercambiar de inmediato (scope exacto)
-    if incoming_code:
-        try:
-            flow = Flow.from_client_secrets_file(
-                CLIENT_SECRET_FILE,
-                scopes=scopes_callback or SCOPES,   # ← usa los scopes de retorno si vienen
-                redirect_uri=REDIRECT_URI,
-            )
-            sess_state = st.session_state.get("oauth_state")
-            if sess_state and incoming_state and incoming_state != sess_state:
-                st.warning("Aviso: 'state' no coincide (posible reinicio de sesión). Continuamos…")
-
-            flow.fetch_token(code=incoming_code)
-            creds = flow.credentials
-            service = build("drive", "v3", credentials=creds)
-
-            st.session_state.service = service
-            st.session_state.pop("oauth_state", None)
-            _clear_qp()  # limpia ?code&state&scope para no reintentar
-
-            st.success("✅ Autenticación correcta (Drive).")
-            return service
-
-        except HttpError as e:
-            st.error(f"Error de Google API: {e}"); st.stop()
-        except Exception as e:
-            st.error(f"Error en Web OAuth (intercambiando code): {e}"); st.stop()
-
-    # 2) Sin code → inicia autorización y muestra URL
-    try:
-        flow = Flow.from_client_secrets_file(
-            CLIENT_SECRET_FILE,
-            scopes=SCOPES,
-            redirect_uri=REDIRECT_URI,
-        )
-        auth_url, state = flow.authorization_url(
-            prompt="consent",
-            access_type="offline",
-            include_granted_scopes="true",  # si te vuelve a mezclar scopes, el paso 1 ya lo maneja
-        )
-        st.session_state["oauth_state"] = state
-
-        st.markdown("### 🔑 Autenticación con Google Drive")
-        st.markdown(f"[Autorizar con Google]({auth_url})")
-        st.code(REDIRECT_URI, language="text")
-        st.info("Tras autorizar, Google te devolverá aquí con ?code=… y seguiremos automáticamente.")
-        st.stop()
-
-    except HttpError as e:
-        st.error(f"Error de Google API: {e}"); st.stop()
-    except Exception as e:
-        st.error(f"Error iniciando Web OAuth: {e}"); st.stop()
-
-
-def listar_bfs_drive(service, root_folder_id):
-    queue = [root_folder_id]
-    files = []
-    while queue:
-        current = queue.pop(0)
-        page_token = None
-        while True:
-            resp = safe_execute(
-                service.files().list(
-                    q=f"'{current}' in parents and trashed=false",
-                    fields="nextPageToken, files(id,name,mimeType,modifiedTime)",
-                    pageSize=1000, pageToken=page_token,
-                    includeItemsFromAllDrives=True, supportsAllDrives=True
-                )
-            )
-            for f in resp.get("files", []):
-                if f["mimeType"] == "application/vnd.google-apps.folder":
-                    queue.append(f["id"])
-                else:
-                    files.append(f)
-            page_token = resp.get("nextPageToken")
-            if not page_token:
-                break
-    return files
-
-def extraer_texto_drive(service, file_id, mime_type):
-    if mime_type == "application/pdf":
-        data = safe_execute(service.files().get_media(fileId=file_id))
-        fh = io.BytesIO(data)
-        reader = PdfReader(fh)
-        return "\n".join([(p.extract_text() or "") for p in reader.pages]).strip()
-    elif mime_type == "application/vnd.google-apps.document":
-        data = safe_execute(service.files().export(fileId=file_id, mimeType="text/plain"))
-        return data.decode("utf-8", errors="ignore")
-    elif mime_type in ("text/plain", "text/markdown"):
-        data = safe_execute(service.files().get_media(fileId=file_id))
-        return data.decode("utf-8", errors="ignore")
-    return None
-
-def get_acl_drive(service, file_id):
-    try:
-        resp = safe_execute(service.permissions().list(
-            fileId=file_id, fields="permissions(id,type,domain)", supportsAllDrives=True
-        ))
-        perms = resp.get("permissions", []) or []
-        return {
-            "permissionIds": [p["id"] for p in perms if p.get("type") == "user" and "id" in p],
-            "domains": [p["domain"].lower() for p in perms if p.get("type") == "domain" and p.get("domain")],
-            "anyone": any(p.get("type") == "anyone" for p in perms),
-        }
-    except Exception:
-        return {"permissionIds": [], "domains": [], "anyone": False}
-
-def get_current_user_drive(service):
-    me = safe_execute(service.about().get(fields="user(emailAddress,permissionId)")) or {}
-    u = me.get("user", {}) or {}
-    email = (u.get("emailAddress") or "").lower()
-    domain = email.split("@")[-1] if "@" in email else ""
-    return {"email": email, "permissionId": u.get("permissionId"), "domain": domain.lower()}
-
-# cache de verificación en vivo (drive)
-def _live_cache_get(key):
-    cache = st.session_state.setdefault("drive_live_check", {})
-    v = cache.get(key)
-    if not v: return None
-    if time.time() - v["ts"] > 300: return None
-    return v["ok"]
-
-def _live_cache_set(key, ok):
-    cache = st.session_state.setdefault("drive_live_check", {})
-    cache[key] = {"ok": bool(ok), "ts": time.time()}
-
-def drive_can_read(service, file_id: str) -> bool:
-    if not file_id: return False
-    ck = f"id:{file_id}"
-    cached = _live_cache_get(ck)
-    if cached is not None: return cached
-    try:
-        safe_execute(service.files().get(fileId=file_id, fields="id", supportsAllDrives=True))
-        _live_cache_set(ck, True); return True
-    except HttpError as e:
-        if getattr(e, "resp", None) and e.resp.status in (403, 404):
-            _live_cache_set(ck, False); return False
-        _live_cache_set(ck, False); return False
-    except Exception:
-        _live_cache_set(ck, False); return False
-
-# ─────────────────────────────────────────────────────────────────────────────
-# DROPBOX
-# ─────────────────────────────────────────────────────────────────────────────
-
-def oauth_dropbox():
-    if not DROPBOX_APP_KEY or not DROPBOX_APP_SECRET:
-        st.error("❌ Falta DROPBOX_APP_KEY o DROPBOX_APP_SECRET en .env")
-        return None
-
-    # Crea el flow una vez y guárdalo (PKCE)
-    if "dbx_auth_flow" not in st.session_state:
-        st.session_state.dbx_auth_flow = dropbox.DropboxOAuth2FlowNoRedirect(
-            consumer_key=DROPBOX_APP_KEY,
-            consumer_secret=DROPBOX_APP_SECRET,
-            token_access_type="offline",
-            scope=["files.metadata.read", "files.content.read"],
-            include_granted_scopes="user",
-            use_pkce=True,
-        )
-        st.session_state.dbx_authorize_url = st.session_state.dbx_auth_flow.start()
-        st.session_state.dbx_code = ""  # buffer del código
-
-    st.markdown("### 🔑 Conectar Dropbox")
-    st.write("1) Abre esta URL, autoriza la app y copia el **código**:")
-    st.code(st.session_state.dbx_authorize_url, language="text")
-
-    # Campo persistente (sin form)
-    st.session_state.dbx_code = st.text_input(
-        "2) Pega aquí el código de autorización de Dropbox",
-        value=st.session_state.get("dbx_code", ""),
-    )
-
-    col1, col2 = st.columns(2)
-    finish = col1.button("Finalizar conexión Dropbox")
-    cancel = col2.button("Cancelar")
-
-    if cancel:
-        for k in ("dbx_auth_flow", "dbx_authorize_url", "dbx_code"):
-            st.session_state.pop(k, None)
-        st.info("Conexión Dropbox cancelada.")
-        return None
-
-    if not finish:
-        return None
-
-    # Limpiar y aceptar también URL completa pegada
-    code = (st.session_state.get("dbx_code") or "").strip()
-    code = code.lstrip(":").strip()
-    if code.startswith("http"):
-        from urllib.parse import urlparse, parse_qs
-        qs = parse_qs(urlparse(code).query)
-        code = (qs.get("code") or [""])[0].strip()
-
-    if not code:
-        st.warning("Pega el código primero.")
-        return None
-
-    try:
-        oauth_result = st.session_state.dbx_auth_flow.finish(code)
-        for k in ("dbx_auth_flow", "dbx_authorize_url", "dbx_code"):
-            st.session_state.pop(k, None)
-
-        dbx = dropbox.Dropbox(
-            oauth2_access_token=oauth_result.access_token,
-            oauth2_refresh_token=oauth_result.refresh_token,
-            app_key=DROPBOX_APP_KEY,
-            app_secret=DROPBOX_APP_SECRET,
-        )
-        _ = dbx.users_get_current_account()
-        st.success("✅ Dropbox conectado")
-        return dbx
-    except Exception as e:
-        st.error(f"Error conectando a Dropbox: {e}")
-        return None
-
-
-def dropbox_list_files(dbx, root_path):
-    root = "" if not root_path or root_path.strip() in ("", "/") else root_path.strip()
-    if root and not root.startswith("/"):
-        root = "/" + root
-
-    files = []
-    try:
-        res = dbx.files_list_folder(root, recursive=True, include_non_downloadable_files=False)
-        while True:
-            for e in res.entries:
-                if isinstance(e, dropbox.files.FileMetadata):
-                    files.append(e)
-            if res.has_more:
-                res = dbx.files_list_folder_continue(res.cursor)
-            else:
-                break
-    except ApiError as e:
-        st.error(f"Error listando Dropbox en '{root}': {e}"); st.stop()
-    return files
-
-
-def _guess_mime_from_name(name: str) -> str:
-    n = (name or "").lower()
-    if n.endswith(".pdf"): return "application/pdf"
-    if n.endswith(".md"):  return "text/markdown"
-    if n.endswith(".txt"): return "text/plain"
-    return "application/octet-stream"
-
-def dropbox_read_text(dbx, file_id=None, path_lower=None, mime=None):
-    path_or_id = file_id or path_lower
-    try:
-        meta, resp = dbx.files_download(path_or_id)
-        data = resp.content
-        if mime == "application/pdf":
-            fh = io.BytesIO(data); reader = PdfReader(fh)
-            return "\n".join([(p.extract_text() or "") for p in reader.pages]).strip()
-        elif mime in ("text/plain", "text/markdown"):
-            return data.decode("utf-8", errors="ignore")
-    except ApiError:
-        return None
-    return None
-
-def dropbox_can_read(dbx, file_id: str) -> bool:
-    try:
-        dbx.files_get_metadata(file_id)
-        return True
-    except ApiError:
-        return False
-    
-# ─────────────────────────────────────────────────────────────────────────────
-# ONEDRIVE (MICROSOFT GRAPH)
-# ─────────────────────────────────────────────────────────────────────────────
-
-def _ms_headers(token_dict):
-    return {"Authorization": f"Bearer {token_dict['access_token']}"}
-
-# ─────────────────────────────────────────────────────────────────────────────
-# ONEDRIVE (MICROSOFT GRAPH) — Device Code Flow sólido y único
-# ─────────────────────────────────────────────────────────────────────────────
-
-def onedrive_device_login():
-    """
-    OneDrive login simplificado (MSAL Device Code), estilo oauth_login_drive:
-    - Un único método con UI mínima.
-    - Muestra enlace de autorización; al pulsar “He autorizado” canjea el código.
-    - Guarda el token en st.session_state.onedrive_token y lo devuelve.
-    - No solicita scopes reservados ('offline_access', 'openid', 'profile').
-    - Usa tenant específico si ONEDRIVE_TENANT_ID está definido; si no, 'common'.
-    """
-    # Validaciones básicas
-    if not ONEDRIVE_CLIENT_ID:
-        st.error("❌ Falta ONEDRIVE_CLIENT_ID en .env"); st.stop()
-
-    # Scopes por defecto y limpieza de reservados
-    default_scopes = ["Files.Read", "User.Read"]
-    scopes_env = (ONEDRIVE_SCOPES or default_scopes)
-    scopes = list({*scopes_env})  # dedup
-    for reserved in ("offline_access", "openid", "profile"):
-        if reserved in scopes:
-            scopes.remove(reserved)
-
-    # Authority (tenant específico o 'common')
-    tenant = (ONEDRIVE_TENANT_ID or "").strip() or "common"
-    authority = f"https://login.microsoftonline.com/{tenant}"
-
-    # Inicia o reutiliza el Device Code flow
-    if "od_flow" not in st.session_state:
-        try:
-            app = msal.PublicClientApplication(client_id=ONEDRIVE_CLIENT_ID, authority=authority)
-            flow = app.initiate_device_flow(scopes=scopes)
-            if "user_code" not in flow:
-                err = flow.get("error_description") or flow.get("error") or "Error iniciando Device Code flow"
-                st.error(f"❌ {err}"); st.stop()
-            st.session_state["od_flow"] = flow
-            st.session_state["od_authority"] = authority
-        except Exception as e:
-            st.error(f"❌ Error MSAL: {e}"); st.stop()
-
-    # UI mínima (idéntico patrón a Drive)
-    flow = st.session_state["od_flow"]
-    verify_url = flow.get("verification_uri_complete") or flow.get("verification_uri") or "https://microsoft.com/devicelogin"
-    user_code = flow.get("user_code", "")
-
-    st.markdown("### 🔐 Conectar OneDrive")
-    st.markdown(f"[Abrir página de autorización]({verify_url})")
-    if "verification_uri_complete" not in flow and user_code:
-        st.write("Si te pide un código, introduce este:")
-        st.code(user_code, language="text")
-
-    col1, col2, col3 = st.columns(3)
-    go     = col1.button("🚀 He autorizado", use_container_width=True, type="primary")
-    reset  = col2.button("🔁 Reiniciar", use_container_width=True)
-    cancel = col3.button("Cancelar", use_container_width=True)
-
-    if reset:
-        st.session_state.pop("od_flow", None)
-        st.session_state.pop("od_authority", None)
-        st.info("Flujo reiniciado. Vuelve a pulsar el botón para generar una nueva URL.")
-        st.stop()
-
-    if cancel:
-        st.session_state.pop("od_flow", None)
-        st.session_state.pop("od_authority", None)
-        st.info("Conexión OneDrive cancelada.")
-        st.stop()
-
-    if not go:
-        st.stop()
-
-    # Canjea y guarda token
-    try:
-        app = msal.PublicClientApplication(client_id=ONEDRIVE_CLIENT_ID, authority=st.session_state["od_authority"])
-        result = app.acquire_token_by_device_flow(flow)
-        if "access_token" not in result:
-            err = result.get("error_description") or result.get("error") or "No se obtuvo access_token"
-            st.error(f"❌ Error en la autorización: {err}")
-            st.session_state.pop("od_flow", None)
-            st.session_state.pop("od_authority", None)
-            st.stop()
-
-        st.session_state.onedrive_token = result
-        st.session_state.pop("od_flow", None)
-        st.session_state.pop("od_authority", None)
-        st.success("✅ OneDrive conectado.")
-        try:
-            st.rerun()
-        except Exception:
-            try:
-                st.experimental_rerun()
-            except Exception:
-                pass
-        return result
-    
-    except Exception as e:
-        st.error(f"❌ Error canjeando el código: {e}")
-        st.session_state.pop("od_flow", None)
-        st.session_state.pop("od_authority", None)
-        st.stop()
-
-
-def onedrive_list_files(token_dict, root_path):
-    """Lista recursivamente archivos (pdf/txt/md) desde root_path ('' o '/subcarpeta')."""
-    headers = _ms_headers(token_dict)
-
-    # ---- PRE-CHEQUEO: ¿tiene OneDrive aprovisionado y permisos? ----
-    probe = requests.get(f"{GRAPH}/me/drive", headers=headers, timeout=20)
-    if probe.status_code != 200:
-        try:
-            err = probe.json().get("error", {})
-            code = err.get("code", "")
-            msg  = err.get("message", "")
-        except Exception:
-            code = ""; msg = probe.text
-        # Mensajes útiles
-        if probe.status_code in (400, 404) and ("OneDrive" in msg or "drive" in msg.lower() or "site" in msg.lower()):
-            raise RuntimeError(
-                "OneDrive del usuario no está aprovisionado. Abre OneDrive una vez en el navegador "
-                "(portal.office.com → OneDrive) para inicializarlo y vuelve a intentarlo."
-            )
-        if probe.status_code in (401, 403):
-            raise RuntimeError(
-                f"Permisos insuficientes en Graph (HTTP {probe.status_code}). "
-                f"Código: {code or 'N/A'} · Mensaje: {msg or 'N/A'}. "
-                "Comprueba que el login concedió 'Files.Read' o 'Files.Read.All' y que, si es un tenant corporativo, el admin dio consentimiento."
-            )
-        raise RuntimeError(f"Graph /me/drive devolvió {probe.status_code}. {code}: {msg}")
-
-    # ---- Funciones auxiliares para listar ----
-    def _children_url(path):
-        if not path or path.strip() in ("", "/"):
-            return f"{GRAPH}/me/drive/root/children?$select=id,name,file,folder,lastModifiedDateTime"
-        p = path if path.startswith("/") else f"/{path}"
-        return f"{GRAPH}/me/drive/root:{p}:/children?$select=id,name,file,folder,lastModifiedDateTime"
-
-    files = []
-
-    def walk(url):
-        while url:
-            r = requests.get(url, headers=headers, timeout=30)
-            if r.status_code == 404:
-                st.warning(f"⚠️ La ruta de OneDrive {root_path!r} no existe. Usando raíz.")
-                return walk(_children_url(""))
-            if r.status_code >= 400:
-                # Saca el detalle del error para ver realmente qué pasa
-                try:
-                    e = r.json().get("error", {})
-                    code = e.get("code", "")
-                    msg  = e.get("message", "")
-                except Exception:
-                    code = ""; msg = r.text
-                raise RuntimeError(f"Graph dijo {r.status_code} {code}: {msg}")
-            data = r.json()
-            for it in data.get("value", []):
-                if it.get("folder"):
-                    sub_url = f"{GRAPH}/me/drive/items/{it['id']}/children?$select=id,name,file,folder,lastModifiedDateTime"
-                    walk(sub_url)
-                elif it.get("file"):
-                    mime = (it["file"].get("mimeType") or "").lower()
-                    files.append({
-                        "id": it["id"],
-                        "name": it["name"],
-                        "mimeType": mime,
-                        "modifiedTime": it.get("lastModifiedDateTime"),
-                    })
-            url = data.get("@odata.nextLink")
-
-    walk(_children_url(root_path))
-    keep = ("application/pdf", "text/plain", "text/markdown")
-    return [f for f in files if (f["mimeType"] in keep or f["name"].lower().endswith((".pdf",".txt",".md")))]
-
-
-def onedrive_download(token_dict, item_id, mime_hint=None):
-    headers = _ms_headers(token_dict)
-    url = f"{GRAPH}/me/drive/items/{item_id}/content"
-    r = requests.get(url, headers=headers, timeout=60)
-    if r.status_code == 403:
-        return None
-    r.raise_for_status()
-    data = r.content
-    mime = (mime_hint or "").lower()
-    if mime == "application/pdf":
-        fh = io.BytesIO(data)
-        reader = PdfReader(fh)
-        return "\n".join([(p.extract_text() or "") for p in reader.pages]).strip()
-    else:
-        try:
-            return data.decode("utf-8", errors="ignore")
-        except Exception:
-            return None
-
-
-def onedrive_can_read(token_dict, item_id: str) -> bool:
-    headers = _ms_headers(token_dict)
-    url = f"{GRAPH}/me/drive/items/{item_id}?$select=id"
-    r = requests.get(url, headers=headers, timeout=20)
-    return r.status_code == 200
-
+import streamlit as st
+
+# Config
+from src.config.config import *
+
+# Connectors
+from src.connectors.drive import drive_can_read, get_current_user_drive, oauth_login_drive, construir_vectorstore_drive
+from src.connectors.dropbox import dropbox_can_read, oauth_dropbox, construir_vectorstore_dropbox
+from src.connectors.onedrive import onedrive_can_read, onedrive_device_login, construir_vectorstore_onedrive
+from src.utility.helpers import cosine_dist
 
 # ─────────────────────────────────────────────────────────────────────────────
 # PERMISOS UNIFICADOS
@@ -698,234 +63,6 @@ def has_access(service_or_token, doc_metadata, user_ctx=None):
     # valida en vivo siempre (por si cambió el ACL o acceso por grupo)
     return drive_can_read(service_or_token, fid) if fid else bool(fast_allow)
 
-
-# ─────────────────────────────────────────────────────────────────────────────
-# CONSTRUCCIÓN ÍNDICES
-# ─────────────────────────────────────────────────────────────────────────────
-
-@st.cache_resource(show_spinner=False)
-def construir_vectorstore_drive_cached(files_serializable, creds_dict, batch_size=200, persist_path="faiss_index"):
-    def load_manifest(path):
-        if os.path.exists(path):
-            try:
-                with open(path, "r", encoding="utf-8") as f:
-                    return json.load(f)
-            except Exception:
-                pass
-        return {"processed_ids": [], "total_chunks": 0, "completed": False, "started_at": datetime.now().isoformat()}
-
-    def save_manifest(path, state):
-        tmp = path + ".tmp"
-        with open(tmp, "w", encoding="utf-8") as f:
-            json.dump(state, f, ensure_ascii=False, indent=2)
-        os.replace(tmp, path)
-
-    os.makedirs(persist_path, exist_ok=True)
-    manifest_path = os.path.join(persist_path, "progress.json")
-    state = load_manifest(manifest_path)
-    processed_ids = set(state.get("processed_ids", []))
-    total_chunks = int(state.get("total_chunks", 0))
-    was_completed = bool(state.get("completed", False))
-
-    embeddings = OpenAIEmbeddings(model="text-embedding-3-small")
-    vectorstore = None
-    try:
-        vectorstore = FAISS.load_local(persist_path, embeddings, allow_dangerous_deserialization=True)
-        print(f"📂 (Drive) Cargando índice desde {persist_path}")
-        if was_completed:
-            print("✅ Índice ya completo. Usando caché.")
-            return vectorstore
-    except Exception:
-        pass
-
-    if vectorstore is None:
-        index = faiss.IndexFlatL2(1536)
-        vectorstore = FAISS(embedding_function=embeddings, index=index, docstore=InMemoryDocstore({}), index_to_docstore_id={})
-
-    service = build("drive", "v3", credentials=Credentials.from_authorized_user_info(creds_dict))
-    splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=100)
-
-    files_serializable = [dict(t) if not isinstance(t, dict) else t for t in files_serializable]
-    remaining_files = [f for f in files_serializable if f["id"] not in processed_ids]
-
-    docs_batch, pending_ids = [], []
-
-    def flush(reason="batch"):
-        nonlocal docs_batch, pending_ids, total_chunks, state
-        if not docs_batch: return
-        vectorstore.add_documents(docs_batch)
-        total_chunks += len(docs_batch)
-        vectorstore.save_local(persist_path)
-        for fid in pending_ids: processed_ids.add(fid)
-        state.update({"processed_ids": list(processed_ids), "total_chunks": total_chunks, "completed": False})
-        save_manifest(manifest_path, state)
-        print(f"🧩 (Drive) Persistidos {len(docs_batch)} chunks [{reason}]")
-        docs_batch, pending_ids = [], []
-
-    for idx, f in enumerate(files_serializable):
-        if f["id"] in processed_ids: continue
-        txt = extraer_texto_drive(service, f["id"], f["mimeType"])
-        if not txt: continue
-        acl = get_acl_drive(service, f["id"])
-        base_doc = Document(
-            page_content=txt,
-            metadata={
-                "source": "drive",
-                "title": f["name"], "id": f["id"], "mimeType": f["mimeType"],
-                "modifiedTime": f.get("modifiedTime"), "acl": acl
-            }
-        )
-        chunks = splitter.split_documents([base_doc])
-        docs_batch.extend(chunks); pending_ids.append(f["id"])
-        if len(docs_batch) >= batch_size: flush("lote")
-
-    if docs_batch: flush("final")
-    if not processed_ids and total_chunks == 0: raise RuntimeError("No se encontraron documentos legibles (Drive).")
-    state["completed"] = True; save_manifest(manifest_path, state)
-    print(f"💾 (Drive) Índice guardado en {persist_path}")
-    return vectorstore
-
-def construir_vectorstore_drive(service):
-    files = [f for f in listar_bfs_drive(service, FOLDER_ID) if f["mimeType"] in (
-        "application/pdf", "application/vnd.google-apps.document", "text/plain", "text/markdown"
-    )]
-    files_small = [{"id": f["id"], "name": f["name"], "mimeType": f["mimeType"], "modifiedTime": f.get("modifiedTime")} for f in files]
-    files_serializable = tuple(tuple(sorted(d.items())) for d in files_small)
-    creds = service._http.credentials
-    creds_dict = {
-        "token": creds.token,
-        "refresh_token": getattr(creds, "refresh_token", None),
-        "token_uri": creds.token_uri,
-        "client_id": creds.client_id,
-        "client_secret": creds.client_secret,
-        "scopes": list(getattr(creds, "scopes", []) or SCOPES),
-    }
-    return construir_vectorstore_drive_cached(files_serializable, creds_dict)
-
-@st.cache_resource(show_spinner=False)
-def construir_vectorstore_dropbox_cached(file_tuples, token, batch_size=200, persist_path="faiss_index_dropbox"):
-    embeddings = OpenAIEmbeddings(model="text-embedding-3-small")
-    os.makedirs(persist_path, exist_ok=True)
-    try:
-        vectordb = FAISS.load_local(persist_path, embeddings, allow_dangerous_deserialization=True)
-        print(f"📂 (Dropbox) Cargando índice desde {persist_path}")
-        return vectordb
-    except Exception:
-        pass
-
-    index = faiss.IndexFlatL2(1536)
-    vectordb = FAISS(embedding_function=embeddings, index=index, docstore=InMemoryDocstore({}), index_to_docstore_id={})
-
-    dbx = dropbox.Dropbox(token)
-    splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=100)
-
-    docs_batch = []
-    def flush():
-        nonlocal docs_batch
-        if not docs_batch: return
-        vectordb.add_documents(docs_batch)
-        vectordb.save_local(persist_path)
-        print(f"🧩 (Dropbox) Persistidos {len(docs_batch)} chunks")
-        docs_batch = []
-
-    for f in file_tuples:
-        fid = f["id"]; name = f["name"]; path_lower = f["path_lower"]; mime = f["mimeType"]
-        if mime not in ("application/pdf", "text/plain", "text/markdown"): continue
-        txt = dropbox_read_text(dbx, file_id=fid, path_lower=path_lower, mime=mime)
-        if not txt: continue
-        base = Document(
-            page_content=txt,
-            metadata={"source":"dropbox","id":fid,"path_lower":path_lower,"title":name,"mimeType":mime}
-        )
-        chunks = splitter.split_documents([base])
-        docs_batch.extend(chunks)
-        if len(docs_batch) >= batch_size: flush()
-    if docs_batch: flush()
-    print(f"💾 (Dropbox) Índice guardado en {persist_path}")
-    return vectordb
-
-def construir_vectorstore_dropbox(dbx):
-    files = dropbox_list_files(dbx, DROPBOX_ROOT or "")
-    files_small = [{"id": f.id, "name": f.name, "path_lower": f.path_lower, "mimeType": _guess_mime_from_name(f.name)} for f in files]
-    files_serializable = tuple(tuple(sorted(d.items())) for d in files_small)
-    return construir_vectorstore_dropbox_cached(files_serializable, dbx._oauth2_access_token)
-
-@st.cache_resource(show_spinner=False)
-def construir_vectorstore_onedrive_cached(file_tuples, token_dict, batch_size=200, persist_path="faiss_index_onedrive"):
-    embeddings = OpenAIEmbeddings(model="text-embedding-3-small")
-    os.makedirs(persist_path, exist_ok=True)
-
-    # ⚠️ Normaliza: tuplas -> dicts
-    file_dicts = [dict(t) if not isinstance(t, dict) else t for t in file_tuples]
-
-    try:
-        vectordb = FAISS.load_local(persist_path, embeddings, allow_dangerous_deserialization=True)
-        print(f"📂 (OneDrive) Cargando índice desde {persist_path}")
-        return vectordb
-    except Exception:
-        pass
-
-    index = faiss.IndexFlatL2(1536)
-    vectordb = FAISS(embedding_function=embeddings, index=index, docstore=InMemoryDocstore({}), index_to_docstore_id={})
-    splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=100)
-
-    docs_batch = []
-    printed = 0
-
-    def flush():
-        nonlocal docs_batch
-        if not docs_batch:
-            return
-        vectordb.add_documents(docs_batch)
-        vectordb.save_local(persist_path)
-        print(f"🧩 (OneDrive) Persistidos {len(docs_batch)} chunks")
-        docs_batch = []
-
-    for f in file_dicts:
-        fid  = f["id"]
-        name = f["name"]
-        mime = (f.get("mimeType") or "").lower()
-
-        # 👇 Log UI + consola de cada documento (no bloquea si no hay contexto Streamlit)
-        try:
-            st.write(f"• Indexando **{name}** ({mime}) …")
-        except Exception:
-            pass
-        print(f"(OneDrive) Indexando: {name} [{mime}] id={fid}")
-
-        txt  = onedrive_download(token_dict, fid, mime_hint=mime)
-        if not txt:
-            try:
-                st.warning(f"Saltado: {name} (vacío o no legible)")
-            except Exception:
-                pass
-            continue
-
-        base = Document(page_content=txt, metadata={"source":"onedrive","id":fid,"title":name,"mimeType":mime})
-        chunks = splitter.split_documents([base])
-        docs_batch.extend(chunks)
-        printed += 1
-
-        if len(docs_batch) >= batch_size:
-            flush()
-
-    if docs_batch:
-        flush()
-    print(f"💾 (OneDrive) Índice guardado en {persist_path}")
-    try:
-        st.success(f"✅ OneDrive indexado: {printed} archivos.")
-    except Exception:
-        pass
-    return vectordb
-
-
-def construir_vectorstore_onedrive(token_dict):
-    files = onedrive_list_files(token_dict, ONEDRIVE_ROOT or "")
-    files_small = [{"id": f["id"], "name": f["name"], "mimeType": f.get("mimeType", "")} for f in files]
-    files_serializable = tuple(tuple(sorted(d.items())) for d in files_small)
-    return construir_vectorstore_onedrive_cached(files_serializable, token_dict)
-
-
 # ─────────────────────────────────────────────────────────────────────────────
 # RESPONDER (multi-origen)
 # ─────────────────────────────────────────────────────────────────────────────
@@ -962,7 +99,7 @@ def responder_multi(query, vectordbs, services, threshold=0.50, k=6, chunk_chars
     q_vec = emb.embed_query(query)
     texts = [(d.page_content or "")[:chunk_chars] for d in allowed_chunks]
     d_vecs = emb.embed_documents(texts)
-    paired = list(zip(allowed_chunks, [_cosine(q_vec, dv) for dv in d_vecs]))
+    paired = list(zip(allowed_chunks, [cosine_dist(q_vec, dv) for dv in d_vecs]))
     paired.sort(key=lambda x: x[1], reverse=True)
 
     picked = [d for d, s in paired if s >= float(threshold)][:k] or [d for d, s in paired[:k]]
@@ -984,6 +121,32 @@ def responder_multi(query, vectordbs, services, threshold=0.50, k=6, chunk_chars
         t = picked[0].metadata.get("title", "(sin título)")
         ans = f"Según [{t}]: {(picked[0].page_content or '')[:chunk_chars]}"
     return ans
+
+# ─────────────────────────────────────────────────────────────────────────────
+# ESTILOS
+# ─────────────────────────────────────────────────────────────────────────────
+
+st.markdown("""
+<style>
+.topbar{display:flex;justify-content:space-between;align-items:center;
+padding:6px 10px;border-bottom:1px solid #eaeaea;margin-bottom:8px;}
+.topbar img{height:40px;}
+.chat-bubble-user{
+  background:#dcf8c6;border-radius:20px;padding:10px 14px;display:inline-block;
+  margin:6px 0;max-width:80%;
+}
+.chat-bubble-bot{
+  background:#ffffff;border:1px solid #eee;border-radius:20px;padding:10px 14px;
+  display:inline-block;margin:6px 0;max-width:80%;
+}
+.chat-row{display:flex;align-items:flex-start;}
+.chat-row.user{justify-content:flex-end;}
+.chat-row.bot{justify-content:flex-start;}
+section[data-testid="stSidebar"] .block-container { display:flex; flex-direction:column; align-items:center; text-align:center; }
+section[data-testid="stSidebar"] img { display:block; margin-left:auto; margin-right:auto; }
+section[data-testid="stSidebar"] .stButton>button { width:100%; max-width:240px; margin:0 auto; }
+</style>
+""", unsafe_allow_html=True)
 
 # ─────────────────────────────────────────────────────────────────────────────
 # SIDEBAR (sesiones y fuentes)
@@ -1119,9 +282,6 @@ if st.session_state.get("onedrive_token") and "vectordb_onedrive" not in st.sess
     except Exception as e:
         st.error(f"❌ No se pudo construir el índice de OneDrive: {e}")
 
-
-
-
 # ─────────────────────────────────────────────────────────────────────────────
 # UI PRINCIPAL (chat)
 # ─────────────────────────────────────────────────────────────────────────────
@@ -1138,6 +298,7 @@ for m in st.session_state.messages:
     st.markdown(f'<div class="chat-row {cls}"><div class="{bub}">{m["content"]}</div></div>', unsafe_allow_html=True)
 
 prompt = st.chat_input("Escribe tu mensaje…")
+
 if prompt:
     st.session_state.messages.append({"role": "user", "content": prompt})
     st.markdown(f'<div class="chat-row user"><div class="chat-bubble-user">{prompt}</div></div>', unsafe_allow_html=True)
